create util function for feature type
